Move datafixture to it's own module
